Add persistence to circuits.
A lot of things have changed. You can either have a circuit without any exercises, or with them.

Do not try and make a circuit or exercise with the constructors; use the new create methods.

More screens will dynamically update now (hopefully). There's still a few cases where the data is only set once, but they're relatively few now and only change infrequently.

On the mobile side, we should try to pass around circuit IDs instead of full circuit objects. This'll result in fewer bugs in serialisation and might be a bit faster.

Six hours was way too long to spend on this one commit.